+++ tom_issue_tracker/test/tom_issue_tracker_test.dart
@@ -1,0 +1,8 @@
+import 'package:tom_issue_tracker/tom_issue_tracker.dart';
+import 'package:test/test.dart';
+
+void main() {
+  test('calculate', () {
+    expect(calculate(), 42);
+  });
+}